refactor: extract helpers to ensnode-core package
