+++ .github/workflows/static-analysis.yml
@@ -25,16 +25,16 @@ jobs:
           node-version-file: .nvmrc
 
       - name: Audit dependencies
-        run: cd apps/ensnode && pnpm audit --audit-level=low
+        run: pnpm audit --audit-level=low
 
       - name: Install dependencies
-        run: cd apps/ensnode && pnpm install --frozen-lockfile
+        run: pnpm install --frozen-lockfile
 
       - name: Run Biome CI
-        run: cd apps/ensnode && pnpm biome ci
+        run: pnpm biome ci
 
       - name: Run TypeScript type checks
-        run: cd apps/ensnode && pnpm run typecheck
+        run: pnpm -r typecheck
 
       - name: Run Ponder build checks
         run: cd apps/ensnode && pnpm codegen
