+++ .github/workflows/unit-tests.yml
@@ -25,7 +25,7 @@ jobs:
           node-version-file: .nvmrc
 
       - name: Install dependencies
-        run: cd apps/ensnode && pnpm install --frozen-lockfile
+        run: pnpm install --frozen-lockfile
 
       - name: Run unit tests
-        run: cd apps/ensnode && pnpm test
+        run: pnpm -r test
