+++ apps/ensnode/package.json
@@ -8,7 +8,7 @@
     "start": "ponder start",
     "db": "ponder db",
     "codegen": "ponder codegen",
-    "test": "vitest",
+    "test": "echo 'no test to run'",
     "lint": "biome check --write .",
     "typecheck": "tsc"
   },
@@ -26,7 +26,6 @@
     "graphql-yoga": "^5.10.9",
     "hono": "^4.6.14",
     "ponder": "^0.8.26",
-    "ts-deepmerge": "^7.0.2",
     "viem": "^2.21.57"
   },
   "devDependencies": {
@@ -37,8 +36,5 @@
   },
   "engines": {
     "node": ">=18.14"
-  },
-  "resolutions": {
-    "vite": "5.4.12"
   }
 }
