+++ apps/ensnode/src/handlers/Registrar.ts
@@ -2,8 +2,12 @@ import { type Context } from "ponder:registry";
 import schema from "ponder:schema";
 import { Block } from "ponder";
 import { type Hex, labelhash, namehash } from "viem";
+import {
+  isLabelIndexable,
+  makeSubnodeNamehash,
+  tokenIdToLabel,
+} from "../../../../packages/ensnode-core/src/subname-helpers";
 import { upsertAccount, upsertRegistration } from "../lib/db-helpers";
-import { isLabelIndexable, makeSubnodeNamehash, tokenIdToLabel } from "../lib/subname-helpers";
 
 const GRACE_PERIOD_SECONDS = 7776000n; // 90 days in seconds
