+++ apps/ensnode/src/plugins/base.eth/ponder.config.ts
@@ -2,8 +2,12 @@ import { type ContractConfig, createConfig, factory } from "ponder";
 import { http } from "viem";
 import { base } from "viem/chains";
 
-import { blockConfig, rpcEndpointUrl, rpcMaxRequestsPerSecond } from "../../lib/helpers";
-import { createPluginNamespace } from "../../lib/plugin-helpers";
+import {
+  blockConfig,
+  rpcEndpointUrl,
+  rpcMaxRequestsPerSecond,
+} from "../../../../../packages/ensnode-core/src/helpers";
+import { createPluginNamespace } from "../../../../../packages/ensnode-core/src/plugin-helpers";
 import { BaseRegistrar } from "./abis/BaseRegistrar";
 import { EarlyAccessRegistrarController } from "./abis/EARegistrarController";
 import { L2Resolver } from "./abis/L2Resolver";
