+++ apps/ensnode/src/plugins/linea.eth/ponder.config.ts
@@ -2,8 +2,12 @@ import { ContractConfig, createConfig, factory, mergeAbis } from "ponder";
 import { http, getAbiItem } from "viem";
 
 import { linea } from "viem/chains";
-import { blockConfig, rpcEndpointUrl, rpcMaxRequestsPerSecond } from "../../lib/helpers";
-import { createPluginNamespace } from "../../lib/plugin-helpers";
+import {
+  blockConfig,
+  rpcEndpointUrl,
+  rpcMaxRequestsPerSecond,
+} from "../../../../../packages/ensnode-core/src/helpers";
+import { createPluginNamespace } from "../../../../../packages/ensnode-core/src/plugin-helpers";
 import { BaseRegistrar } from "./abis/BaseRegistrar";
 import { EthRegistrarController } from "./abis/EthRegistrarController";
 import { NameWrapper } from "./abis/NameWrapper";
