+++ packages/ensnode-core/src/helpers.ts
